Cleaned up user agent code (#14)
Cleaned up user agent code
